@@ -12,26 +12,25 @@ declare(strict_types=1);
 
 namespace App\Traits;
 
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Support\Str;
-
 /**
  * Trait TableSorting
  * Provides functionalities to handle table sorting and filtering in a structured manner.
  */
 trait TableSorting
 {
-    public $sortField;
+    public ?string $sortField = null;
 
     public $columName;
 
-    public $filtervalue;
+    public mixed $filterValue = null;
 
-    public $sortDirection = false;
+    public $clickColumn;
+
+    public bool $sortDirection = true;
 
     public $nameRelashion;
 
-    public $listFilterValues;
+    public $modelName;
 
     /**
      * Orders the query results by the specified column.
@@ -42,60 +41,14 @@ trait TableSorting
      *
      * @param  string  $columValue  The name of the column to order by.
      */
-    public function orderColumBy(string $columValue): void
+    public function orderColumBy(string $column): void
     {
-        if ($this->columName === $columValue) {
+        if ($this->sortField === $column) {
             $this->sortDirection = ! $this->sortDirection;
         } else {
-            $this->sortDirection = true;
+            $this->sortField = $column;
+            $this->sortDirection = true; // default ASC on new column
         }
-        $this->columName = $columValue;
-    }
-
-    public function makeQueryByColumn($query)
-    {
-        return $this->getQueryValues() === null ? $query : $this->buildQuery($query);
-
-    }
-
-    /**
-     * Builds the query with ordering based on the current column and sort direction.
-     *
-     * This method orders the query results by the column name retrieved from `getQueryValues()`
-     * and the current sort direction. The sort direction is determined by the `sortDirection` property.
-     *
-     * @param  Builder  $query  The query builder instance.
-     * @return Builder The modified query builder instance.
-     */
-    public function buildQuery($query)
-    {
-        return $query->orderBy($this->getQueryValues(), $this->sortDirection ? 'asc' : 'desc');
-    }
-
-    /**
-     * Filters the query results based on the provided search value.
-     *
-     * This method modifies the query to filter results based on the specified search value.
-     * If the filter value contains '_id', it calls a relationship method to handle the query.
-     * Otherwise, it applies a 'like' filter to the query and orders the results.
-     *
-     * @param  string  $stringvalue  The value to search for.
-     * @param  Builder  $query  The query builder instance.
-     * @return Builder The modified query builder instance.
-     */
-    public function makeQueryBySearch(string $stringvalue, $query)
-    {
-        $stringvalue = mb_strtolower(trim($stringvalue));
-
-        if (Str::contains($this->filtervalue, '_id')) {
-
-            return $query->{$this->nameRelashion}($stringvalue, $this->filtervalue);
-
-        }
-
-        return $query->whereRaw('LOWER('.$this->filtervalue.') like ?', ['%'.$stringvalue.'%'])
-            ->orderBy($this->filtervalue, $this->sortDirection ? 'asc' : 'desc');
-
     }
 
     /**
@@ -106,31 +59,37 @@ trait TableSorting
      *
      * @param  string  $modelName  The name of the model to initialize table sorting for.
      */
-    public function inicializteTableSorting(string $modelName): void
+    public function setupTableSorting(string $modelName): void
     {
-        $model = "\\App\\Models\\$modelName";
-
+        $this->modelName = "\\App\\Models\\$modelName";
         $this->resetPage();
         $this->resetOrdersValues();
-        $this->filtervalue = $model::$startFilterBay;
-        $this->listFilterValues = $model::getFilterableAttributes();
+        $this->filterValue = $this->getModelDefaultFilter($modelName);
     }
 
     /**
-     * Retrieves the current column name used for sorting.
+     * Gets the default filter field for a model.
+     * First tries to get from config, then falls back to interface method.
      *
-     * This method returns the column name that is currently set for sorting.
-     * If the column name is not set, it returns null.
-     *
-     * @return string|null The column name used for sorting, or null if not set.
+     * @param  string  $modelName  The name of the model.
+     * @return string The default filter field.
      */
-    protected function getQueryValues(): ?string
+    private function getModelDefaultFilter(string $modelName): string
     {
-        if (empty($this->columName)) {
-            return null;
+        $defaultFilter = config("model_filters.default_filters.{$modelName}");
+
+        if ($defaultFilter) {
+            return $defaultFilter;
         }
 
-        return $this->columName;
+        // Fallback to interface method if model implements Filterable
+        $modelClass = "\\App\\Models\\$modelName";
+        if (method_exists($modelClass, 'getDefaultFilterField')) {
+            return $modelClass::getDefaultFilterField();
+        }
+
+        // Ultimate fallback
+        return 'id';
     }
 
     /**
@@ -141,8 +100,9 @@ trait TableSorting
      */
     private function resetOrdersValues(): void
     {
+        $this->clickColumn = null;
         $this->columName = null;
-        $this->sortDirection = false;
+        $this->sortDirection = true;
         $this->sortField = null;
     }
 }
