@@ -19,18 +19,20 @@ final class Credential extends Model
         'credential_code',
     ];
 
-    public static function credentialExist($idcredential, $credentialnumbre, $credentialModelId = null): bool
+    public static function credentialExist($idcredential, $credentialNumber, $credentialModelId = null): bool
     {
         if ($credentialModelId) {
 
-            return self::whereHas('medicals', static function ($query) use ($credentialnumbre, $credentialModelId) {
-                $query->where('credential_number', $credentialnumbre)
-                    ->where('medical_id', '!=', $credentialModelId);
-            })->where('id', $idcredential)->exists();
+            return self::where('id', $idcredential)
+                ->whereHas('medicals', function ($query) use ($credentialNumber, $credentialModelId) {
+                    $query->where('credential_number', $credentialNumber)
+                        ->whereNotIn('medicals.person_id', [$credentialModelId]);
+
+                })->exists();
         }
 
-        return self::whereHas('medicals', static function ($query) use ($credentialnumbre) {
-            $query->where('credential_number', $credentialnumbre);
+        return self::whereHas('medicals', static function ($query) use ($credentialNumber) {
+            $query->where('credential_number', $credentialNumber);
         })->where('id', $idcredential)->exists();
     }
 
@@ -43,7 +45,7 @@ final class Credential extends Model
     protected function credentialName(): Attribute
     {
         return Attribute::make(
-            set: fn ($value) => ucwords(mb_strtolower(trim($value))),
+            set: fn ($value) => ucwords(mb_strtolower(mb_trim($value))),
 
         );
     }
@@ -51,7 +53,7 @@ final class Credential extends Model
     protected function credentialCode(): Attribute
     {
         return Attribute::make(
-            set: fn ($value) => mb_strtoupper(mb_strtolower(trim((string) $value))),
+            set: fn ($value) => mb_strtoupper(mb_strtolower(mb_trim((string) $value))),
 
         );
     }
