@@ -4,30 +4,40 @@ declare(strict_types=1);
 
 namespace App\Models;
 
+use App\Interfaces\Filterable;
 use Database\Factories\PatientFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
-final class Patient extends Model
+final class Patient extends Model implements Filterable
 {
     /** @use HasFactory<PatientFactory> */
     use HasFactory;
 
     protected $fillable = [
-        'person_id',
-        'patient_address', 'patient_photo'];
+        'person_id', 'blood_type_id', 'patient_photo', 'patient_weight', 'patient_height'];
+
+    protected $casts = [
+        'person_id' => 'integer',
+        'patient_photo' => 'string',
+    ];
+
+    public static function getDefaultFilterField(): string
+    {
+        return 'person_name';
+    }
+
+    public static function getRelationModel(): array
+    {
+        return [
+            'person:*',
+            'person.document:id,document_name',
+        ];
+    }
 
     public function person(): BelongsTo
     {
         return $this->belongsTo(Person::class);
-    }
-
-    protected function casts(): array
-    {
-        return [
-            'person_id' => 'integer',
-            'patient_datebirth' => 'date',
-        ];
     }
 }
