@@ -21,7 +21,7 @@ final class Specialty extends Model
     protected function specialtyName(): Attribute
     {
         return Attribute::make(
-            set: fn ($value) => ucwords(mb_strtolower(trim($value))),
+            set: fn ($value) => ucwords(mb_strtolower(mb_trim($value))),
 
         );
     }
