@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use App\Models\Credential;
 use App\Models\Degree;
+use App\Models\Person;
 use App\Models\Specialty;
 use App\Models\State;
 use Illuminate\Database\Migrations\Migration;
@@ -19,16 +20,11 @@ return new class extends Migration
     {
         Schema::create('medicals', function (Blueprint $table) {
             $table->id();
+            $table->foreignIdFor(Person::class)->constrained();
             $table->foreignIdFor(State::class)->constrained();
             $table->foreignIdFor(Credential::class)->constrained();
             $table->foreignIdFor(Specialty::class)->nullable()->constrained();
             $table->foreignIdFor(Degree::class)->nullable()->constrained();
-            $table->string('medical_name');
-            $table->string('medical_lastname');
-            $table->string('medical_address')->nullable();
-            $table->string('medical_phone')->nullable();
-            $table->string('medical_email')->nullable();
-            $table->string('medical_dni')->nullable()->unique();
             $table->timestamps();
         });
     }
