@@ -1,7 +1,7 @@
 <th
     {{$attributes}}
     scope="col"
-    class="px-2 py-2 text-left text-sm font-normal text-gray-500 rtl:text-right dark:text-gray-400"
+    class="px-1.5 py-2 text-left text-sm font-normal text-gray-500 rtl:text-right dark:text-gray-400"
 >
     {{ $slot }}
 </th>
