@@ -1,6 +1,6 @@
 <x-mail::message>
 # Saludos.
-Les informamos que hemos registrado exitosamente la actualización de los datos de su sucursal en nuestro sistema.
+Les informamos que hemos registrado exitosamente la actualización de los datos de la sucursal en nuestro sistema.
 
 Contar con información precisa y actualizada nos permite optimizar los procesos de comunicación y garantizar una mejor gestión administrativa.
 
